+++ Sources/VaniCore/Audio/AudioCapturing.swift
@@ -1,5 +1,12 @@
 public protocol AudioCapturing: Sendable {
   func start() async throws
   func stop() async throws -> CapturedAudio
+  func recoverPendingAudio() async throws -> CapturedAudio?
   func cancel() async
+}
+
+extension AudioCapturing {
+  public func recoverPendingAudio() async throws -> CapturedAudio? {
+    nil
+  }
 }